+++ src/ode/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod ekf;
+pub(crate) mod explicit_rk;
 pub mod predictions;
+pub(crate) mod rosenbrock;
 pub mod solver;
 
 pub use predictions::*;